move reachability into matching
