+++ proto/internal/temporal/server/api/matchingservice/v1/request_response.proto
@@ -256,9 +256,10 @@ message GetWorkerVersioningRulesRequest {
     string namespace_id = 1;
     string task_queue = 2;
     temporal.api.enums.v1.TaskQueueType task_queue_type = 3;
+    google.protobuf.Duration deleted_rule_inclusion_period = 4;
 
     oneof command {
-        temporal.api.workflowservice.v1.GetWorkerVersioningRulesRequest request = 4;
+        temporal.api.workflowservice.v1.GetWorkerVersioningRulesRequest request = 5;
     }
 }
